Moved OCI8Conn into connection.go
